requirements.txt
tensorflow~=2.10.1
numpy~=1.24.4
six~=1.15.0
sentencepiece~=0.2.0
seqeval~=1.2.2
sacrebleu~=2.2.0
scipy~=1.10.1
scikit-learn~=1.3.2
pyyaml~=5.4.1
opencv-python~=4.9.0.80
pillow~=10.3.0
pycocotools~=2.0.7
matplotlib~=3.7.5
immutabledict~=4.2.0
tqdm~=4.66.2
pandas~=2.0.3
requests~=2.31.0
setuptools~=69.5.1
h5py~=3.1.0
tensorboard~=2.10.1
contextlib2~=21.6.0
keras~=2.10.0
lvis~=0.5.3
lxml~=5.2.1
mediapipe~=0.10.11